cinny: remove the build environment from the build results
For some reason cinny included the entire build environment in the JS
bundle. This seems odd and the author says they don't recall why that
was required other than for some bug.

Thankfully we don't build this in an environment that contains
priviledged information (everything is sandboxed with nix; yay!) so
there are no AWS/GCP/... credentials to be stolen from us :-)

Conversation in the cinny channel:
 - https://matrix.to/#/!rgQbWTfsoUcbvEDAvs:matrix.org/$DB3YVlF9JJrtRewkawZXbX3qQBcqo_BqziQECrHZ-KQ?via=matrix.org&via=envs.net&via=maunium.net
 - https://matrix.to/#/!rgQbWTfsoUcbvEDAvs:matrix.org/$cSG6HMYgs0YC5bKVlq3N6ODQ_iIo4xY-k0Erd7NRpuM?via=matrix.org&via=envs.net&via=maunium.net